+++ src/judgement.cpp
@@ -46,7 +46,7 @@ namespace judgement {
                 ext_type = ext_t::CXX;
             else
                 ext_type = ext_t::Other;
-            return input_t{name, ext, ext_type, std::stoi(id)};
+            return input_t{name, ext, ext_type, std::stol(id)};
         }
     }
